Changed the function to work with floats.
This is needed for future super-sampling implementation as this is the place
where the variability will be introduced.

+++ src/renderer/renderer.h
@@ -15,9 +15,9 @@ namespace Svit
 			render (World& _world, Settings& _settings, Engine& _engine) = 0;
 
 			float 
-			compute_normalized_coordinate (int value, int max_value)
+			compute_normalized_coordinate (float value, float max_value)
 			{
-				return ((((float)value + 0.5f)/(float)max_value)*2.0f)-1.0f;
+				return (((value + 0.5f)/max_value)*2.0f)-1.0f;
 			}
 	};
 }
